% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enrollment_line_chart.R
\name{mod_enrollment_line_chart_ui}
\alias{mod_enrollment_line_chart_ui}
\title{enrollment_line_chart UI Function}
\usage{
mod_enrollment_line_chart_ui(id)
}
\arguments{
\item{id}{A unique identifier, linking the UI to the Server}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_enrollment_line_chart_ui("enrollment_line_chart_1")

To be copied in the server
mod_enrollment_line_chart_server("enrollment_line_chart_1")'
}
