% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst_diagram.R
\name{mod_sunburst_diagram_server}
\alias{mod_sunburst_diagram_server}
\title{sunburst_diagram Server Functions}
\usage{
mod_sunburst_diagram_server(
  id,
  df,
  step_cols = c("entity_category_1", "entity_category_2", "entity_category_3",
    "entity_outcome")
)
}
\arguments{
\item{id}{Identifier to link server and UI for the module.}

\item{df}{A data-frame.}

\item{step_cols}{Which columns of the data-frame should define the rings in the sunburst
diagram.}
}
\description{
To be copied in the UI
mod_sunburst_diagram_ui("sunburst_diagram_1")
}
\details{
To be copied in the server
mod_sunburst_diagram_server("sunburst_diagram_1")
}
