% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst_diagram.R
\name{mod_sunburst_diagram_ui}
\alias{mod_sunburst_diagram_ui}
\title{sunburst_diagram UI Function}
\usage{
mod_sunburst_diagram_ui(id, title, subtitle = NULL)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{title, subtitle}{Title and subtitle for the page/module.}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_sunburst_diagram_ui("sunburst_diagram_1")

To be copied in the server
mod_sunburst_diagram_server("sunburst_diagram_1")
}
